Restructure code for user info rework
