+++ jupyterhub_singleuser_profiles/profiles.py
@@ -5,53 +5,34 @@ import json
 import logging
 import re
 from kubernetes.client import V1EnvVar, V1ResourceRequirements, V1ConfigMap, V1ObjectMeta, V1SecurityContext, V1Capabilities, V1SELinuxOptions, V1Secret, V1Volume, V1VolumeMount, V1PersistentVolumeClaimVolumeSource
-import base64
-from kubernetes.client.rest import ApiException
-from openshift.dynamic import DynamicClient
+
 from .service import Service
 from .utils import escape, parse_resources
 from .sizes import Sizes
 from .images import Images
+from .openshift import OpenShift
+from .user import User
 
 _LOGGER = logging.getLogger(__name__)
 
 _JUPYTERHUB_USER_NAME_ENV = "JUPYTERHUB_USER_NAME"
-_USER_CONFIG_MAP_TEMPLATE = "jupyterhub-singleuser-profile-%s"
 _USER_CONFIG_PROFILE_NAME = "@singleuser@"
 _GPU_KEY = "nvidia.com/gpu"
-_DEFAULT_USER_CM = {
-  "env":[],
-  "gpu":"0",
-  "last_selected_image":"",
-  "last_selected_size":"",
-}
-_DEFAULT_USER_SECRET = {
-  "env":[],
-}
 
 class SingleuserProfiles(object):
   GPU_MODE_SELINUX = "selinux"
   GPU_MODE_PRIVILEGED = "privileged"
   def __init__(self, namespace=None, verify_ssl=True, gpu_mode=None, service_account_path='/var/run/secrets/kubernetes.io/serviceaccount'):
     self.profiles = []
-    self.api_client = None
     self.namespace = namespace #TODO why do I need to pass namespace?
     self.gpu_mode = gpu_mode
-    self.oapi_client = None
 
-    if not self.namespace:
-      with open(os.path.join(service_account_path, 'namespace')) as fp:
-          self.namespace = fp.read().strip()
-    kubernetes.config.load_incluster_config()
-    self.api_client = kubernetes.client.CoreV1Api()
+    self.openshift = OpenShift(namespace=namespace, verify_ssl=verify_ssl)
 
-    configuration = kubernetes.client.Configuration()
-    configuration.verify_ssl = verify_ssl
-    self.oapi_client = DynamicClient(
-      kubernetes.client.ApiClient(configuration=configuration)
-    )
+    self.service = Service(self.openshift, self.namespace)
+    self.images = Images(self.openshift, namespace=namespace)
+    self.user = User(self.openshift, default_image=self.images.get_default())
 
-    self.service = Service(self.oapi_client, self.namespace)
   @property
   def gpu_mode(self):
     return self._gpu_mode
@@ -64,142 +45,15 @@ class SingleuserProfiles(object):
       self._gpu_mode = self.GPU_MODE_SELINUX
     else:
       self._gpu_mode = None
-
-  def get_config_maps_matching_label(self, target_label='jupyterhub=singleuser-profiles'):
-    config_maps_list = []
-    try:
-      config_maps = self.api_client.list_namespaced_config_map(self.namespace, label_selector=target_label)
-    except ApiException as e:
-      if e.status != 404:
-        _LOGGER.error(e)
-      return config_maps_list
-    for cm in config_maps.items:
-      config_maps_list.append(cm.metadata.name)
-
-    _LOGGER.info("Found these additional Config Maps: %s" % config_maps_list)
-    return config_maps_list
-
-  def read_config_map(self, config_map_name, key_name="profile"):
-    try:
-      config_map = self.api_client.read_namespaced_config_map(config_map_name, self.namespace)
-    except ApiException as e:
-      if e.status != 404:
-        _LOGGER.error(e)
-      return {}
-      
-    config_map_yaml = yaml.load(config_map.data[key_name])
-    return config_map_yaml
-
-  def read_secret(self, secret_name, key_name="profile"):
-    try:
-      secret = self.api_client.read_namespaced_secret(secret_name, self.namespace)
-    except ApiException as e:
-      if e.status != 404:
-        _LOGGER.error(e)
-      return {}
-    if secret.data:
-      base64_secret = secret.data[key_name]
-      base64_bytes = base64_secret.encode('ascii')
-      secret_bytes = base64.b64decode(base64_bytes)
-      content = secret_bytes.decode('ascii')
-      secret_yaml = yaml.load(content)
-    else:
-      secret_yaml = {}
-    return secret_yaml
-
-  def write_config_map(self, config_map_name, key_name, data):
-    cm = V1ConfigMap()
-    cm.metadata = V1ObjectMeta(name=config_map_name, labels={'app': 'jupyterhub'})
-    cm.data = {key_name: yaml.dump(data, default_flow_style=False)}
-    try: 
-      api_response = self.api_client.replace_namespaced_config_map(config_map_name, self.namespace, cm)
-    except ApiException as e:
-      if e.status == 404:
-        try: 
-          api_response = self.api_client.create_namespaced_config_map(self.namespace, cm)
-        except ApiException as e:
-          _LOGGER.error("Exception when calling CoreV1Api->create_namespaced_config_map: %s\n" % e)
-      else:
-        raise
-
-  def write_secret(self, secret_name, key_name, data):
-    secret = V1Secret()
-    secret.metadata = V1ObjectMeta(name=secret_name, labels={'app': 'jupyterhub'})
-    secret.string_data = {key_name: yaml.dump(data, default_flow_style=False)} #stringData instead of data here to make kubernetes parse this as a string without base64 encoding
-    try:
-      api_response = self.api_client.replace_namespaced_secret(secret_name, self.namespace, secret)
-    except ApiException as e:
-      if e.status == 404:
-        try:
-          api_response = self.api_client.create_namespaced_secret(self.namespace, secret)
-        except ApiException as e:
-          _LOGGER.error("Exception when calling CoreV1Api->create_namespaced_secret: %s\n" % e)
-      else:
-        raise
-
-  def update_user_profile_cm(self, username, data={}):
-    
-    user_cm = self.get_user_profile_cm(username)
-    user_cm_env = []
-    user_secret = self.get_user_secret(username)
-    user_secret['env'] = []
-    resource_name = _USER_CONFIG_MAP_TEMPLATE % escape(username)
-    cm_key_name = "profile"
-    for key, value in data.items():
-      user_cm[key] = value
-    for var in user_cm['env']:
-      if var['type'] == "password":
-        user_secret['env'].append(var) 
-      else:
-        user_cm_env.append(var)
-    user_cm['env'] = user_cm_env
-    self.write_secret(resource_name, cm_key_name, user_secret)
-    self.write_config_map(resource_name, cm_key_name, user_cm)
-
-  def fix_legacy_user_cm(self, cm):
-    env_list = []
-    for key, value in cm['env'].items():
-      env_list.append({"name":key, "type":"text", "value":value})
-    cm['env'] = env_list
-    return cm
-    
-  def get_default_image(self):
-    image_list = self.get_images()
-    if len(image_list) > 0:
-      return image_list[0]
-    return ''
-
-  def get_user_profile_cm(self, username, for_k8s=False):
-    cm = self.read_config_map(_USER_CONFIG_MAP_TEMPLATE % escape(username), "profile")
-    secret = self.get_user_secret(username)
-    if cm == {}:
-      cm = _DEFAULT_USER_CM
-    if isinstance(cm['env'], dict):
-      cm = self.fix_legacy_user_cm(cm)
-    for sec in secret['env']:
-      cm['env'].append(sec)
-    if for_k8s:
-      for x in cm['env']:
-        del x['type'] 
-    if cm['last_selected_image'] == '':
-      cm['last_selected_image'] = self.get_default_image()
-    return cm
-
-  def get_user_secret(self, username):
-    secret = self.read_secret(_USER_CONFIG_MAP_TEMPLATE % escape(username))
-    if secret == {}:
-      return _DEFAULT_USER_SECRET
-    else:
-      return secret 
   
   def load_profiles(self, secret_name="jupyter-singleuser-profiles", filename=None, key_name="jupyterhub-singleuser-profiles.yaml", username=None):
     self.profiles = []
     self.sizes = []
-    if self.api_client:
+    if self.openshift.api_client:
       profiles_config_maps = [secret_name]
-      profiles_config_maps.extend(sorted(self.get_config_maps_matching_label()))
+      profiles_config_maps.extend(sorted(self.openshift.get_config_maps_matching_label()))
       for cm_name in profiles_config_maps:
-        config_map_yaml = self.read_config_map(cm_name, key_name)
+        config_map_yaml = self.openshift.read_config_map(cm_name, key_name)
         if config_map_yaml:
           self.sizes.extend(config_map_yaml.get("sizes", [self.empty_profile()]))
           self.profiles.extend(config_map_yaml.get("profiles", [self.empty_profile()]))
@@ -208,7 +62,7 @@ class SingleuserProfiles(object):
       if len(self.profiles) == 0:
         self.profiles.append(self.empty_profile())
       if username:
-        config_map_yaml = self.get_user_profile_cm(username, for_k8s=True)
+        config_map_yaml = self.user.get(username, for_k8s=True)
         if config_map_yaml:
           if not config_map_yaml.get('name'):
             config_map_yaml['name'] = _USER_CONFIG_PROFILE_NAME
@@ -288,18 +142,6 @@ class SingleuserProfiles(object):
     
   def get_sizes(self):
     return self.sizes
-
-  def get_images(self):
-    i = Images(self.oapi_client, self.namespace)
-    return i.get_form(name_only=True)
-  
-  def get_image_list_form(self, username=None):
-
-    if not self.profiles:
-      self.load_profiles(username=username)
-
-    i = Images(self.oapi_client, self.namespace)
-    return i.get_form(self.get_profile_by_name(_USER_CONFIG_PROFILE_NAME).get('last_selected_image'))
 
   @classmethod
   def empty_profile(self):
